rag-app/requirements.txt
# RAG Learning Application - Requirements

# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.12.0

# Vector Store and Search
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0

# Web Interface
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
requests>=2.31.0

# Utilities
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
plotly>=5.10.0

# Optional: For better performance
# accelerate>=0.20.0
# bitsandbytes>=0.41.0
